requirements.txt
opencv_python==4.3.0.36
matplotlib==3.1.3
torch==1.5.1
tqdm==4.46.0
pydicom==2.0.0
numpy==1.18.1
torchvision==0.6.1
Pillow==8.0.1
